backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
pdfplumber==0.10.2
python-docx==0.8.11
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
nltk==3.8.1
werkzeug==2.3.6
gunicorn==20.1.0